refactor: restructure dirs to use only kebab case
